integrating neutrino and much name change fixes

+++ chain/fullblocktests/params.go
@@ -107,9 +107,9 @@ var regressionNetParams = &chaincfg.Params{
 	BIP0065Height:            1351,      // Used by regression tests
 	BIP0066Height:            1251,      // Used by regression tests
 	SubsidyReductionInterval: 150,
-	TargetTimespan:           time.Hour * 24 * 14, // 14 days
-	TargetTimePerBlock:       time.Minute * 10,    // 10 minutes
-	RetargetAdjustmentFactor: 4,                   // 25% less, 400% more
+	TargetTimespan:           int64(time.Hour * 24 * 14), // 14 days
+	TargetTimePerBlock:       int64(time.Minute * 10),    // 10 minutes
+	RetargetAdjustmentFactor: 4,                          // 25% less, 400% more
 	ReduceMinDifficulty:      true,
 	MinDiffReductionTime:     time.Minute * 20, // TargetTimePerBlock * 2
 	GenerateSupported:        true,
